updates27-6-2021

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserRequest;
 use App\Http\Requests\UserUpdateRequest;
+use App\Http\Resources\UsersTrashedResource;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
 use App\Models\Admin;
 use App\Http\Resources\UsersResource;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use App\Traits\GeneralTrait;
 
@@ -37,8 +39,42 @@ class UserController extends Controller
             $offset = $request->start;
         }
 
-        $list = Admin::orderByDesc('created_at')->offset($offset)->take($perPage)->get();
+        $list = Admin::withoutTrashed()->orderByDesc('created_at')->offset($offset)->take($perPage)->get();
         $arr = UsersResource::collection($list);
+        $recordsTotal = Admin::get()->count();
+        $recordsFiltered = Admin::get()->count();
+        return response()->json([
+            'draw' => $request->draw,
+            'recordsTotal' => $recordsTotal,
+            'recordsFiltered' => $recordsFiltered,
+            'data' => $arr
+        ]);
+    }
+
+    /////////////////////////////////////////
+    /// Get Trashed users index
+    public function trashedUser()
+    {
+        $title = trans('menu.trashed_users');
+        return view('admin.users.trashed-users', compact('title'));
+    }
+
+    /////////////////////////////////////////
+    /// Get Trashed users
+    public function getTrashedUsers(Request $request)
+    {
+        $perPage = 10;
+        if ($request->has('length')) {
+            $perPage = $request->length;
+        }
+
+        $offset = 0;
+        if ($request->has('start')) {
+            $offset = $request->start;
+        }
+
+        $list = Admin::onlyTrashed()->orderByDesc('created_at')->offset($offset)->take($perPage)->get();
+        $arr = UsersTrashedResource::collection($list);
         $recordsTotal = Admin::get()->count();
         $recordsFiltered = Admin::get()->count();
         return response()->json([
@@ -139,7 +175,20 @@ class UserController extends Controller
         return $this->returnSuccessMessage(trans('general.update_success_message'));
 
     }
+    /////////////////////////////////////////
+    /// User restore
+    public function restore(Request $request)
+    {
+        if ($request->ajax()) {
+            $user = Admin::onlyTrashed()->find($request->id);
+            if (!$user) {
+                return redirect()->route('admin.not.found');
+            }
+            $user->restore();
+            return $this->returnSuccessMessage(trans('general.restore_success_message'));
+        }
 
+    }
     /////////////////////////////////////////
     /// User Destroy
     public function destroy(Request $request)
@@ -149,10 +198,26 @@ class UserController extends Controller
             if (!$user) {
                 return redirect()->route('admin.not.found');
             }
+            $user->delete();
+
+            return $this->returnSuccessMessage(trans('general.delete_success_message'));
+        }
+
+    }
+
+    /////////////////////////////////////////
+    /// user force delete
+    public function forceDelete(Request $request)
+    {
+        if ($request->ajax()) {
+            $user = Admin::onlyTrashed()->find($request->id);
+            if (!$user) {
+                return redirect()->route('admin.not.found');
+            }
             if (!empty($user->photo)) {
                 Storage::delete($user->photo);
             }
-            $user->delete();
+            $user->forceDelete();
 
             return $this->returnSuccessMessage(trans('general.delete_success_message'));
         }
